Refactor the error checker utility into its own library.
